Code reorganization.
Move stuff for App and EditTaskDialog windows into separate directories.
Move main-view.jsx into views/
Move TaskList.jsx into components.

TaskList is a component.
MainView is a view (kind of a fragment) to be used.
App and EditTaskDialog are meant to be shown in their own BrowserWindow.
These have an *-index.jsx and a jsx root component for these windows.