Fix bug causing incorrect total time reports.
If the user hits pause, and then stop, getTotalTimeRan() gets called
twice, returning the same value each time.

The same situation exists when the timer expires, and the user hits
stop.

Resetting totalTimeRan to zero when getTotalTimeRan() eliminates this
bug.  We don't need to preserve totalTimeRan when it is consumed by
TaskRunningView.

Generally the task is updated after we get this information from the
Timer.