Times on this form should be in minutes.
We use seconds as the smallest unit of time, but the edit settings form
should work with minutes.  THe user should expect to see minutes, and be
able to enter minutes.

So when we load from settings, convert seconds to minutes, then when the
form is submitted, convert from minutes to seconds.

This keep everyone happy.