Separate updateTask into two functions.
Separate into updateTaskTimeSpentOnTask and taskDone.

Calling updateTask to update the task with timeSpentOnTask resets done
status by setting done to false.

This was an assumption made when creating the application, we were
assuming that when the user clicks on 'stop' the task is *not* done.

But since the app allows users to 'start' a task that is 'done', and add
more time to it, this assumption does not hold.

We should consider not allowing the user to start a task that is 'done'.
Users should be able to mark a task as not 'done' in the task editor.