Don't start tasks that are done.
Don't let the user start tasks that are done.
The user will need to go to EditTaskView, and mark the task as not
'done'.
Then they can start that task.
If the user needs to start a task that is done, that task is obviously
not done.

We may want to consider changing the behavior to making the task not
'done' when the user hits start..  This might be a bit more seamless.