Disable Save button when name is empty string.
When the user first opens the Add Task View, the Save button is enabled.
The name and description are set to the empty string.
We do not want the user to be able to click the Save button in this
case.

The software will not allow form data to be submitted.
But the Save button should not be enabled, to avoid making the user
think that he/she can submit an incomplete form.  We don't want the user
to feel foolish.