Handle Add New Task case.
Default id was -1.  Change this to some uuidv4.

Also reload data from the database after an edit.
We may want to avoid reloads, and just add the new task to our internal
data structures directly.  We may need to split the function into two,
editTask and addNewTask, and handle the two cases separately.