filterTasks, filterTasksByDone, getTasks
These functions use find() under the hood.

filterTasks is the public facing function that takes a filterName, and
dtermines what query to run based on the filterName.

filterTasksByDone does find with selector done.  This gets all the tasks
that are 'done' or 'not done'

getTasks does find with selector type === 'task' This gets all the
documents with the type of 'task'.  We need this because this database
treats indexes as separate documents.  They're called design documents.