Call init() in color() and strobe().
This forces a check to see if the luxafor device is connected to the
system.

This avoids an issue when the luxafor device does not listen to commands
after it's been disconnected / reconnected.

In the above case, init() will force the device to re-initialize, so we
can continue sending commands to the luxafor device.