Move new Luxafor() statement to c'tor.
We don't need to keep creating a new Luxafor object every time we call
init().  Originally init() was meant to be called once, but we now call
it when color() or strobe() are invoked to ensure that the Luxafor flag
is initalized properly.

We may want to rename init() to something else here eventually.