UpdateTaskTimeSpentOnTask should take task id.
Passing UpdateTaskTimeSpentOnTask the task id avoids the case where this
function is being called when TaskRunningView is being unmounted.

In this case, the state transition function will be called first,
setting currentTask to an empty string.

Then the component gets unmounted when App re-renders itself after state
transition.

This causes an issue with looking up the Current Task.

Thus we pass in the task's id so it can be looked up by id, avoiding a
bug in this case.